requirements.txt
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
cssselect==1.1.0
idna==2.10
Jinja2==2.11.2
lxml==4.6.1
MarkupSafe==1.1.1
pdfminer.six==20201018
pdfquery==0.4.3
pycparser==2.20
pycryptodome==3.9.9
pyquery==1.4.1
python-dateutil==2.8.1
requests==2.24.0
roman==3.3
six==1.15.0
sortedcontainers==2.2.2
urllib3==1.25.11
